Updated `catalogsearch` to ignore cases of input catalogs

+++ src/lksearch/MASTSearch.py
@@ -1119,6 +1119,28 @@ class MASTSearch(object):
         # check to see if a cloud_uri exists, if so we just pass that
 
         download = True
+        if not conf.CHECK_CACHED_FILE_SIZES:
+            # If this configuration parameter is set and the file exists
+            # in the cache, we do not search for it
+            local_path = "/".join(
+                [
+                    config.get_cache_dir(),
+                    "mastDownload",
+                    row["obs_collection"],
+                    row["obs_id"],
+                    row["productFilename"],
+                ]
+            )
+            if os.path.isfile(local_path):
+                manifest = pd.DataFrame(
+                    {
+                        "Local Path": [local_path],
+                        "Status": ["UNKNOWN"],
+                        "Message": [None],
+                        "URL": [None],
+                    }
+                )
+                return manifest
         if not conf.DOWNLOAD_CLOUD:
             if pd.notna(row["cloud_uri"]):
                 download = False
@@ -1196,18 +1218,19 @@ class MASTSearch(object):
         ]
 
         manifest = pd.concat(manifest)
-        status = manifest["Status"] != "COMPLETE"
-        if np.any(status):
-            warnings.warn(
-                "Not All Files Downloaded Successfully, Check Returned Manifest.",
-                SearchWarning,
-            )
-            if remove_incomplete:
-                for file in manifest.loc[status]["Local Path"].values:
-                    if os.path.isfile(file):
-                        os.remove(file)
-                        warnings.warn(f"Removed {file}", SearchWarning)
-                    else:
-                        warnings.warn(f"Not a file: {file}", SearchWarning)
-        manifest = manifest.reset_index(drop=True)
+        if conf.CHECK_CACHED_FILE_SIZES:
+            status = manifest["Status"] != "COMPLETE"
+            if np.any(status):
+                warnings.warn(
+                    "Not All Files Downloaded Successfully, Check Returned Manifest.",
+                    SearchWarning,
+                )
+                if remove_incomplete:
+                    for file in manifest.loc[status]["Local Path"].values:
+                        if os.path.isfile(file):
+                            os.remove(file)
+                            warnings.warn(f"Removed {file}", SearchWarning)
+                        else:
+                            warnings.warn(f"Not a file: {file}", SearchWarning)
+            manifest = manifest.reset_index(drop=True)
         return manifest
